add initial composer mode support
Added -composer-root command-line argument which specifies
a composer project root directory that is expected to contain
composer.json file as well as all its sources and vendor dir.

A vendor dir should be generated before kphp2cpp is invoked:

	composer install

After that the $composer-root/vendor should contain all information we need.

Instead of doing a dependency analysis starting from the $composer-root/composer.json
and following the "requires" and "requires-dev" we do a vendor folder traversal.

The benefit is that it's compatible with composer plugins that may change
a way the composer handles composer.json files so it may affect the vendor
directory contents. If we rely on a properly generated vendor folder
we don't need to care about plugins at all.

Refs