add comments for the generated vertex types
Add "comment" optional field in vertex-desc.json for
vertices that may benefit from a little summary.

That comment is added to the docummented vertex type
as a C++ comment.

So, "comment": "lhs() and rhs()" for op_log_and_let will produce
something like this:

	// lhs() and rhs()
	template<>
	class vertex_inner<op_log_and_let> : public vertex_inner<meta_op_binary> {

Another example:

	// a simple var, like $foo; variable name is stored in str_val
	template<>
	class vertex_inner<op_var> final: public vertex_inner<meta_op_base> {

The <name>() notation in most comments refer to the named getter
of the vertex. In the example above, lhs() and rhs() are
accessors from meta_op_binary.

This improves readability of both places: json manifest and
generated C++ code (where we end up after "goto definition").