Improve Exceptions support (#65)
Summary:

* Added support for multiple `catch` clauses per `try` block
* Added `\Throwable` interface; `\Exception` now implements `\Throwable`
* Added `\Error` class that implements `\Throwable` + SPL Exception classes
* Old KPHP code will generate the same C++ code for the try/catch blocks
* All code generated code diff is `new Exception(...)` => `__exception_set_location(new Exception(...), __FILE__, __LINE__)` and an instance cast of exception var if it's not `Throwable`
* Code that catches all possible exceptions does not propagate any (exceptions are tracked)
* This PR does not implement `finally` on purpose

## Exceptions inheritance.

* updated CFG and other passes so they handle the program flow correctly

* codegen generates 2 new cases for the try/catch ops.
  If exception class has derived classes, is_a() is used;
  If exception class has no derived classes, hash comparison is used;
  When the code catches \Exception class, same codegen is used.

Exceptions in CFG propagate like this: if try block catches
all exceptions, we consider that nothing propagates.
Otherwise, we consider all exceptions that can be created inside try block do propagate.
We may do exceptions tracking here as well, but it doesn't affect
the code generation right now.

## PreprocessException pass.

Add `__FILE__` and `__LINE__` arguments to the Exception-derived
construction call in the PreprocessException pass instead of
doing it in the lexer.

The problem with lexer is that it tries to know about the
Exception class and adds these arguments too early.

If Exception class can be extended, then we don't know
whether a `new T()` expression constructs an exception class.
The PreprocessException has class data info, so it can
perform this operation with more accuracy.

We also change the Exception class constructor signature so
it's compatible with the PHP version. It doesn't accept the
file and line arguments now; to set these fields, a new builtin function
is added that is inserted by the compiler.