make `strict_types=1` disable automatic cast insertion for a current file
In PHP, strict_types=1 would break the code at the run time if given
argument types mismatch expected types (expressed via type hints).

Previously, in KPHP it would silently cast these types and we'll
get inconsistent behavior between PHP and KPHP.

This change makes KPHP code a little bit closer to PHP,
but instead of failing at the run time we enable more type checking
for cast params in `strict_types=1` files.

A lot of existing PHP code is not prepared for this, so
this behavior is activated only if `declare()` has a phpdoc
with `@kphp-strict-types-enabled`.

As in PHP, strict_types is checked at the call site.