allow `T->regexp` implicit casts even with `strict_types=1`
With `strict_types=1`, there will be no implicit `string->regexp` and `mixed->regexp`
conversions, but in fact we don't need them as there are various overloadings
that can work with all such types (string, regexp, mixed).

With this change, it will be possible to use string-typed arguments
to `preg_*` functions when `declare(strict_types=1)` is used.