ensure that args evaluation order is left-to-right
Generate a C++ code that enforces left-to-right arguments
evaluation order.

We achieve that by moving arguments evaluation to a
separate statements.

At the very least, this increases the generated C++ code size
and will make it less readable.

To minimize that effect, we apply some tricks.

If a function call contains less that 2 order-dependent argument,
it can be left as is, since the potential evaluation order differences
will be minor.

Right now we define order-dependent argument as an argument
that contains a function call which is not pure.
We could also consider operations that can potentially throw
or produce a warning (like array indexing) as order-dependent,
but it looks like it's quite safe to start from functions only.

The general replacement looks like this:

	$result = f(g1(), g2());
	=>
	$result = ({
		auto &&_arg1 = g1();
		auto &&_arg2 = g2();
		f(_arg1, _arg2);
	});

To reduce the noise from that, there are macros for function
calls with a small number of arguments:

	$result = f(g1(), g2());
	=>
	CALL_ORDERED2(f, g1(), g2());