enable instanceof smart casts for unions
Teach extract_instance_from_type_hint() how to handle union types.
The algorithm used is close to the one that is used during the type inference.

We try to find a common base type for all union variants.
If there are multiple candidates, we'll give up.

This is good enough for most sane class hierarchies.

Fixes #217