added optimization for (v)sprintf call with simple format (#250)
(v)sprintf calls with a simple format string are now converted to concatenation.

A simple format string is a string that only contains the `%s` and `%d` specifiers and 
any other text. If the string contains something else, then such a call will be left as it is.

Example:

echo sprintf("Hello %s", $name);
// converted to:
echo "Hello " . $name;

Depending on the length of the string and count specifiers, the speed of concatenation 
is several times faster. Even for the given example, the speed of `sprintf` is 4 times slower 
than that of concatenation.