rewrite ClassData::get_instance_method() (#226)
With the old approach, we'll get a compile-time error.

The new approach prefers non-abstract methods; it'll traverse the classes and
try to find the first non-abstract method. If it's not possible, it will
return the first abstract method (like we did before). The code is slightly
more complicated than that just to avoid redundant work.

Fixes #215