make it possible to split functions.txt into separate files (#296)
Just as before, we specify a path to a **root** `functions.txt` file.
That file can make a `require_once()` call to another file that
will be processed as a part of builtin declarations.

We achieve this behavior by marking the root functions file with
a flag `is_from_functions_file`. Then, every file required from
it will inherit that flag. `file->is_builtin()` will respect that flag.

This change is motivated not only by a need to logically separate the
definitions, but also to make it possible to move some declarations
to another namespace. For example, FFI-related `CData` class should
be declared in `FFI` namespace (`\FFI\CData`).
This is required to make our implementation (#295) closer to the PHP implementation.