added `is_match` flag to `op_switch`
In PHP, a regular `switch` uses non-strict comparison, which leads
to type casts, but the `match` uses strict comparison, which is why
we need a new `is_match` flag in the `op_switch` node.

If this flag is set, then strict comparison will be used when
generating the code for the `switch`.