allowed 'catch' without variable
A 'catch' without a variable is a catch with the 'without_variable'
flag set with a variable with an autogenerated name. This variable
is further used as a placeholder.

At the stage of code generation, if the 'without_variable' flag is set,
then we simply destroy the current exception, thereby handling it.

The gain in this case is in the absence of moving the current exception
into a temp variable, which, according to benchmarks, however, gives an
increase in speed within the margin of error for simple Exception.