added support for explicit octal literals (0o345)
Currently, delimited numbers are tokens with special types,
while 0o123 is not a valid C++ notation for an octal number,
so it should be changed during parsing.

From this, it turns out that in order to correctly handle
such situations, we either need to introduce two new types of
tokens, octal and octal with separators, or for each octal
number, check for separators (which means a complete pass of
all such numbers in n).