namespace name is interpreted as a single token
KPHP already interprets namespace names as a single token, like
PHP 8. The only change here is that now keywords after the
'namespace' will not be converted to keyword tokens, thus it will
be possible to create a namespace with a name like a keyword,
as in PHP 8.

For example:

namespace fn;