implemented `FFI::string()` function
PHP has 2 forms of that call:

1. With explicit length
2. Without explicit length

For (2) it requires the CData ptr to be `char*` or `const char*` (checked).
It also expects it to be null-terminated (unchecked).
If passed CData would be something incompatible, it gives a run time error.

KPHP would give a compile-time error for (2) function form.

For (1) both PHP and KPHP won't perform any type checking.