fixed string::try_to_float_as_php8 function
Fixed behavior when an empty string was considered a valid number, which led
to incorrect runtime warnings about the difference in behavior between PHP 7
and 8.

Added a new type of test @kphp_runtime_should_not_warn for tests that need to
make sure that the code in it does not generate warnings at runtime.