rewrite template functions and lambdas — completely (#321)
In this huge PR I completely drop the current implementation of @kphp-template functions and lambdas (anonymous functions) — and replace it with a new one, which is coupled with type deduction, preliminary auto-casting, template arguments reification, assumptions refactoring and much more.

All these changes are related. Theoretically, they can be split in two or three interconnected PRs, but it would become even more complicated.

Features and changes:
* lambdas inside lambdas inside template functions etc., no depth limit
* capturing other lambdas in uses, in template functions also
* php lambdas passed to built-in functions are codegenerated as C++ lambdas with no extra classes
* template functions logically and syntactically now seem and behave like generics
* assumptions calculation becomes much more logical; the "not instance" state was dropped, assumptions are calculated on demand traversing the function body only up to this exact place (not traversing the whole function body, as before)
* works noticeably faster than master

See the PR for detailed multi-chaptered explanation.