[compiler] assign autogenerated names for unnamed ffi func args
Previously, we used empty string as a param name.
If there are multiple unnamed params, we'll end up with duplicated
param names with potentially different type hints.

This patch assigns autogenerated names, like `_unnamed_param1`
to these parameters, avoiding the collision error.

Also enabled one of the FFI tests from pointers;
it was intended to be enabled, but perhaps I forgot
to add it to the python script previously (?)

Fixes #424