Resumable-friendly external database drivers embedding API & PDO::MySQL resumable basic version (#393)
The main goal of the PR is to introduce internal resumable-friendly API for embedding asynchronous database drivers (MySQL, PostgreSQL, Redis etc.)

Resumable-friendly means that any network communication based on this API will be considered as a suspendable point by KPHP resumable functions (a.k.a. coroutines). To achieve this the API deeply interacts with KPHP web server network internals and epoll reactor.