add memory_get_allocations builtin KPHP function
memory_get_allocations returns a tuple of `(num_allocations, memory_allocated)`.

The benefit of this function is that it can be used to
measure the allocations between two execution points:
how many allocations happened and how much memory we allocated.

	[$old_num_allocs, $old_mem_allocated] = memory_get_allocations();
	do_something();
	[$new_num_allocs, $new_mem_allocated] = memory_get_allocations();
	$num_allocs = $new_num_allocs - $old_num_allocs;
	$mem_allocated = $new_mem_allocated - $old_mem_allocated;

This is especially useful in benchmarking (see ktest bench).

Since it returns a tuple instead of array,
it doesn't do any heap allocations on its own