[compiler,runtime] generate func call for some concatenations
When compiling a string build op where all operands are string-typed
and number of arguments is no more than 5, generate a function call.

	$x . $y => str_concat($x, $y)
	$x . $y . $z => str_concat($x, $y, $z)
	// and so on up to 5 arguments

This affects the binary size positively: a function call takes up
less machine code bytes. The overall binary size decrease is in
0.6-1% range for concatenation heavy programs with good amount
of statically string-typed operands.

We only support up to 5 arguments because other combinations
are less likely.

Contrary to what could be expected, this "redundant" function call
instead of a hardcoded concatenation gives a performance boost,
not a degradation. Firstly, we always compile runtime with -O3.
This is usually not the case with translated PHP.
Right now the default is `-Os`, but even with `-O2` we can
see that a function call is less expensive than the amount
of optimizations we get for these individual runtime functions.

Also, this patch adds a new optimization borrowed from the Go runtime.
When we have a 2-operand concat and either of the operands is an
empty string, we return other (potentially non-empty) string as a result
instead of allocating a new string. Some experiments show that around
17.6% of 2-operand string concatenations have at least one empty string
operand during the run time.

For this optimizations to be valid, we rely on a fact that if string
is shared and modified, it will be copied just like array.

Benchmark results:

    name                     old time/op    new time/op    delta
    Concat::EmptyEmpty          215ns ± 1%      18ns ± 0%   -91.63%
    Concat::EmptyShort          133ns ± 0%      17ns ± 0%   -87.22%
    Concat::EmptyLong           183ns ± 0%      19ns ± 3%   -89.40%
    Concat::ShortShort         92.3ns ± 1%    63.7ns ± 2%   -30.99%
    Concat::ShortConst         80.5ns ± 2%    62.8ns ± 2%   -22.05%
    Concat::ShortShortShort    94.6ns ± 1%    64.0ns ± 0%   -32.35%
    Concat::ShortConstShort    92.5ns ± 3%    72.7ns ± 2%   -21.44%
    Concat::LongLong            185ns ± 1%     118ns ± 4%   -36.59%
    [Geo mean]                  126ns           44ns        -65.40%

    name                     old alloc/op   new alloc/op   delta
    Concat::EmptyEmpty          16.0B ± 0%      0.0B       -100.00%
    Concat::EmptyShort          24.0B ± 0%      0.0B       -100.00%
    Concat::EmptyLong          1.02kB ± 0%    0.00kB       -100.00%
    Concat::ShortShort          32.0B ± 0%     32.0B ± 0%         ~
    Concat::ShortConst          32.0B ± 0%     32.0B ± 0%         ~
    Concat::ShortShortShort     40.0B ± 0%     40.0B ± 0%         ~
    Concat::ShortConstShort     32.0B ± 0%     32.0B ± 0%         ~
    Concat::LongLong           2.02kB ± 0%    2.02kB ± 0%         ~

    name                     old allocs/op  new allocs/op  delta
    Concat::EmptyEmpty           1.00 ± 0%      0.00       -100.00%
    Concat::EmptyShort           1.00 ± 0%      0.00       -100.00%
    Concat::EmptyLong            1.00 ± 0%      0.00       -100.00%
    Concat::ShortShort           1.00 ± 0%      1.00 ± 0%         ~
    Concat::ShortConst           1.00 ± 0%      1.00 ± 0%         ~
    Concat::ShortShortShort      1.00 ± 0%      1.00 ± 0%         ~
    Concat::ShortConstShort      1.00 ± 0%      1.00 ± 0%         ~
    Concat::LongLong             1.00 ± 0%      1.00 ± 0%         ~