[compiler] add hrtime() and microtime() specializations
This is a port of #251 PR without more problematic regexp
function specializations and refactoring.
preg functions will be ported separately as they require
more unrelated changes to be applied.

In addition to microtime, this patch includes hrtime which
is now generally available as KPHP min PHP version is 7.4 now.

	microtime() -> microtime_string()
	microtime(false) -> microtime_string()
	microtime(true) -> microtime_float()

	hrtime() -> hrtime_array()
	hrtime(false) -> hrtime_array()
	hrtime(true) -> hrtime_int()

The main idea is to get less mixed-typed values in the code
that uses functions those return value type depends on its arguments.