[runtime] make shutdown funcs called even after a timeout
The most challenging part is that timeout is catched in a signal
handler of the running coroutine (run_context) which then
changes the state of the worker and switches the context to exit_context.
Running the shutdown handlers inside the run_context before switching it
is dangerous as arbitrary PHP code is not guaranteed to be safe for
the signal handling context.

If we run these shutdown handlers inside of the worker context,
we may suffer from the `exit()` calls done inside of the shutdown
functions that may cause us to attempt switching the context again
and entering a `finished` state instead of the `error` state we were in.

As a solution to this problem, we use a setjmp+longjmp to implement
our recovery from the `exit()` when running shutdown functions.

We only call shutdown handlers for timeout errors for now.
Executing shutdown handlers should not be a problem for some error
kinds, but some of them are more challenging (like memory limit errors).

Fixes #504