[runtime] move 0xc0+0x80 utf8 to a separate method + add a comment
Also used a while loop instead of `do while`.
The extra offset+1 is performed at the call site, since it
has nothing to do with utf8 subsequent bytes.

This is a part of many patches that we need to apply to make
regexp code at least a little bit more maintainable.