[runtime] fix argv[0] handling for server `-o` and cli modes
Previously, we were adding `argv[0]` twice when KPHP runs
a server in once mode (`-o`) or when we're running a cli script.

Once of the solutions is to re-organize the way we parse command-line
arguments so there is never such a situation where adding something
excessive is never an option. Another approach is to memorize and/or
recognize when we don't need to add this default argument anymore.
This PR implements the latter approach as it requires less changes
and fixed the issue.

Fixes #533