Force common runtime template types to be instantiated in runtime sources (#537)
By default all template types are implicitly instantiated, i.e. the code with substituted types is placed in each translation unit that use the corresponding template. It leads to duplication of common instantiations so many times.

To get rid of this unnecessary duplicates in this PR explicit template instantiation is applied for common array<T> and optional<T> instantiation patterns.

This not only reduces the output binary size, but also forces all such instantiations to be compiled with -O3 that potentially can improve performance in some cases.