[runtime] run shutdown handlers for null iface object method calls
When null interface method is called, the script ends abnormally.
This disrupts the normal script termination process and shutdown
handlers will never be executed.

Since calling a method on a null object is not a fatal error inside
our runtime (we can still continue running some other KPHP code),
we can and should do something like a normal `f$exit(1)` and allow
the shutdown handlers to be executed.

To do that, a new runtime function is introduced: `soft_critical_error()`.
The only difference is that it calls `f$exit(1)` instead of doing
some signal raising business.

We may want to use this new function in other places, where recoverable
critical errors are preferred to the current criticals.
For now, it's only used in one place.

It's also debatable whether we should switch to non-fatal critical errors
in the future, so user-generated critical error do not bypass the
shutdown handlers (I'm not sure users expect that behavior anyway).

Fixes #569