support Modulite inside kphp (everything except composer) (#582)
Modulite is a conception of isolating parts of code
PHP, as a language, allows calling any public class/function from any other place in the code. 
For huge PHP projects, it's bad practice, because if you develop some kernel db connectors and public API, you expect your users to use this API, but you have no guarantee that nobody calls your internals directly, even if you expose a Composer package. PHP has no "modules" in a language — and that's what Modulite would take care about.

The concept of colored functions is more flexible, but Modulite is more generalized. It's designed in a way to be integrated into IDE.

I am merging this, despite it's ready only partially — we are going to start using Modulite in VK, so we need its experimental support in production. It's disabled by default and can be enabled by KPHP_ENABLE_MODULITE env.