[compiler] re-order class fields to reduce objects size
Some classes can benefit from fields reordering to
occupy less space due to the unnecessary padding.

Suppose that we have a class like this:

	class Example {
		public int $a; // size=8
		public bool $b; // size=1
		public int $c; // size=8
		public tuple(bool, bool, bool, bool) $d; // size=4
	}

It's layout in memory would look like this:

	[ 4 ] ref_cnt
	[ 4 ] ! padding
	[ 8 ] a
	[ 1 ] b
	[ 7 ] ! padding
	[ 8 ] c
	[ 4 ] d
	[ 4 ] ! padding (trailing padding due to struct alignment)

The total size would be 40.

With this patch, the compiler will re-arrange some fields to minimize
the amount of padding we would get.

	[ 4 ] ref_cnt
	[ 4 ] d
	[ 8 ] a
	[ 8 ] c
	[ 1 ] b
	[ 7 ] ! padding (trailing padding due to struct alignment)

The result size will be 32.

The algorithm is quite simple:

1. Sort all fields by their size (take alignment into the account)
2. Fill the gap after the `ref_cnt` carefully, if possible
3. Output all remaining fields from the biggest to the smallest

Most of the time, this simple algorithm leads to the optimal layout.
It's tested to never make objects bigger; although it may not
always succeed in making them smaller (the size may remain identical).

The testing results on real world applications show that:

* About 13% of classes can be optimized (their size will be reduced)
* On average, this optimization reduces the size by 12-20%
* Depending on the code, this optimization reduces the memory usage by 6-12%