[compiler] add RestrictionIsset to indexing funcs
If `array_last_value` or `array_first_value` is called
on empty array, the same difference like with normal
array indexing will occur: `0 vs null`, `"" vs null`, etc.

This ability to mark a function as "indexing" makes it
possible to add optimized operations that do indexing as
a part of their operation without losing compiler checks
that only worked with `op_index` previously.

For example:

	$x = explode($sep, $s)[0];
	=>
	$x = explode_nth($sep, $s, 0);

If we rewrite the code like that, we'll lose the isset
restriction tracking unless we mark `explode_nth` as indexing func.