[runtime] add count=1 special case for `implode()` function
Some stats from a real-world app:

    ~0.7% calls with count=0
    ~56%  calls with count=1
    ~43%  calls with count>1

The count=0 case can be optimized by returning the empty string right away,
but since it's so rare, this change is not applied.
Our allocator knows how to allocate an empty-sized string efficiently.

The count=1 case can be optimized by returning the sole array element.
This may require a to_string conversion, so we're wrapping it in `f$strval`.
Since this is a frequent case, this optimization will be applied.

This optimization is quite common in languages with immutable (or COW) strings:

    https://cs.opensource.google/go/go/+/refs/tags/go1.19.2:src/strings/strings.go;l=435

Benchmark results:

    name               old time/op    new time/op    delta
    Implode::Implode1    58.0ns ± 0%    24.9ns ± 4%   -57.07%  (p=0.000 n=8+10)
    Implode::Implode3    84.1ns ± 3%    85.6ns ± 1%    +1.78%  (p=0.021 n=10+10)
    [Geo mean]           69.8ns         46.2ns        -33.90%

    name               old alloc/op   new alloc/op   delta
    Implode::Implode1     32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
    Implode::Implode3     56.0B ± 0%     56.0B ± 0%         ~  (all equal)

    name               old allocs/op  new allocs/op  delta
    Implode::Implode1      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
    Implode::Implode3      1.00 ± 0%      1.00 ± 0%         ~  (all equal)

The case for count!=1 gets extra 2-3ns execution time due to the extra size check.
The main benefit of this change is heap allocations reduction.