[vkext] Fix RPC TL `Maybe<array>` storing (#683)
In typed serialization mode TL type `Maybe %(Vector string)` corresponds PHP type `string[]|null`.
But if you tried to store `["asdf", "asdf"]`, you would get the following error: "Unknown constructor asdf for type Maybe"
So it mistakenly was expected to be like in untyped mode, i.e.: `["_" => "resultTrue", "result" => ["asdf", "asdf"]]`.

It happened because typed mode was determined by the condition like `type of serialized object is not array` in case of Maybe storing. But this condition is not hold in case of `Maybe<Vector>` or `Maybe<Tuple>`.

Now we will use special `typed_mode` flag to distinguish typed and untyped storing mode. We will set it at every typed storing entrypoint and reset after that.