[runtime] make implode faster for vector-like arrays
This patch adds two optimizations to implode implementation:

1. For string-typed vector-like arrays, build the result
   into the result string directly

2. Additionally, use a faster path for empty delimiters

Some stats:

* Most of the implode usages have vector-like arrays as an input (~95%)
* About 3-7% of these cases use empty delimiter

With our default algorithm, we copy all characters (bytes) into the
static buffer, a temporary string. Then we create a new string,
copy all bytes in there and return it as a result.

We can improve that if we can calculate the exact result size fast.
Both of the aspects are important: it should be fast and precise.

For instance, it's hard to calculate the precise size for an array
of integers. It's trivial to calculate the precise size for a string.

Therefore, we're only doing this optimization for string arrays for now.

To make the iteration itself faster, we use a fact that it's enough
to check for `is_vector()` only once and avoid any extra branching
for every iterator increment. This reduces the constant iteration overhead.

The benchmarks below show that we become faster for string-typed vectors.
We don't get any significant performance penalties in other cases.
The benefits are bigger with longer strings (because this optimization
is about doing less byte copying).

The empty delimiter implode speedup is self-explanatory: do less work
per each iteration => profit.

    name                              old time/op    new time/op    delta
    Implode::EmptyArray                 17.3ns ± 4%    16.3ns ±10%        ~
    Implode::Vector1                    20.0ns ± 0%    20.7ns ± 8%        ~
    Implode::Vector3                    87.9ns ± 3%    84.2ns ± 2%        ~
    Implode::Vector16                    292ns ± 0%     272ns ± 1%   -6.80%
    Implode::Vector128                  2.15µs ± 0%    2.03µs ± 0%   -5.77%
    Implode::Vector1024                 17.6µs ± 1%    15.5µs ± 1%  -11.67%
    Implode::Vector16long               1.57µs ± 0%    1.06µs ± 1%  -32.32%
    Implode::Vector128long              21.6µs ± 0%    11.3µs ± 1%  -47.74%
    Implode::Vector16short               329ns ± 0%     306ns ± 0%   -7.01%
    Implode::Vector128short             2.43µs ± 1%    2.25µs ± 1%   -7.47%
    Implode::Map3                       82.0ns ± 0%    85.0ns ± 0%        ~
    Implode::Map16long                   284ns ± 4%     285ns ± 0%        ~
    Implode::MixedVector16               401ns ± 0%     392ns ± 1%        ~
    Implode::IntVector16                 274ns ± 0%     271ns ± 3%        ~
    Implode::IntVector128               2.13µs ± 1%    2.03µs ± 2%        ~
    Implode::FloatVector16              14.5µs ± 1%    14.7µs ± 0%        ~
    Implode::FloatVector128              120µs ± 1%     121µs ± 1%        ~
    Implode::Vector3emptyDelim          85.0ns ± 0%    77.5ns ± 2%   -8.86%
    Implode::Vector16emptyDelim          293ns ± 0%     172ns ± 1%  -41.26%
    Implode::Vector128emptyDelim        2.16µs ± 2%    1.12µs ± 0%  -48.10%
    Implode::Vector1024emptyDelim       17.4µs ± 1%     8.8µs ± 0%  -49.67%
    Implode::MixedVector16emptyDelim     412ns ± 1%     392ns ± 0%   -5.00%
    [Geo mean]                           857ns          729ns       -14.88%