fix tmp string optimization in expr-split contexts
In some cases compiler splits a single expression into
a sequence of statements, like in the example below:

	$s .= substr($s, ...);

It will be compiled as:

	SAFE_SET_FUNC_OP($s, append, substr($s, ...), type)

Which expands to:

	({
		auto tmp = substr($s, ...);
		append(a, tmp);
	})

It effectively splits a single expression into 2.
This may break tmp_string optimizations if expression inside
substr create a new string that has no references outside
of the expression itself. When expression will be evaluated,
that new string will be deallocated, making the tmp_string
invalid (it will outlive it's data source).

To avoid that, introduce a new TmpExpr concept to the
vertex compiler. At the point of the potential expr split,
use that type to avoid issues with tmp strings.
It will materialize strings and make sure that tmp_string
is converted to a proper string that can live for whatever
time is necessary.

To avoid pessimizations in the string concat code,
make str_concat functions work with tmp_strings
too by introducing a str_concat_arg helper type.

There are no significant performance changes after this patch.