Correction of checking the file descriptor mode in the stream_option_setter() function
In the f$curl_setopt() function with options for stream descriptors, it is important to check the opening mode of these descriptors.
This is only possible by accessing a static variable from files.cpp. In order to solve the access problem, a new get_stream_mode() function has been added to files.h and cpp.

The stream_option_setter() function has been fixed:
	- it is allowed to pass null as an argument for unsetting variables;
	- checking for the file descriptor opening mode is now performed via get_stream_mode();
	- the set_option_safe() call has been moved to the beginning of the function to save the error_code field in the easy_context