[k2] pass const coroutine parameters by const ref where it's safe (#1427)
In certain scenarios, it is safe for coroutines to accept arguments by reference. This commit allows the code generator to produce user-defined coroutines with reference parameters. The rationale behind this decision is that all await points are encapsulated within runtime functions, which are implemented with safety in mind.

This commit also updates several runtime functions to accept arguments by const &. An additional // SAFETY comment has been added to explain why this change is considered safe.